+++ actionview/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Allow to pass extra tag attributes to `auto_discovery_link_tag` helper
+
+    *Max Trubetskoy*
+
 *   `translate` should accept nils as members of the `:default`
     parameter without raising a translation missing error.  Fixes a
     regression introduced 362557e.